refactor: split usage-view-helpers.js into shared + per-tab modules
Closes #630

usage-view-helpers.js had grown to 302 lines (just past the 300-line
threshold corrected previously in #357 / #262). Splits the file into:

- usage-view-shared.js     : constants, pure DOM helpers, runMetricCards
                             and createRunBasedTabConfig factory, typedefs,
                             plus the _internals export consumed by tests
- usage-view-agents-tab.js : agentTabConfig + its row/header helpers
- usage-view-tokens-tab.js : tokenTabConfig + its row/card helpers
- usage-view-flows-tab.js  : flowTabConfig + its row/header/name helpers
- usage-view-helpers.js    : minimal public entry (~20 lines) that
                             re-exports TABS/createSection/_internals
                             from shared and wires getTabConfig to the
                             three tab modules

No behaviour change. All 415 unit tests pass (vitest run).
