+++ src/utils/usage-view-helpers.js
@@ -1,302 +1,20 @@
 /**
- * Pure constants and helpers extracted from UsageView.
- * No side-effect dependencies — safe to unit-test in isolation.
- */
-
-import { _el } from './dom-core.js';
-import { formatDuration, formatTokens, runTooltip, rateColor, rateCls } from './usage-formatters.js';
-
-// --- Tab definitions ---
-
-export const TABS = [
-  { id: 'agents', label: 'Agents (Work)' },
-  { id: 'tokens', label: 'Tokens' },
-  { id: 'flows', label: 'Flows' },
-];
-
-// --- Chart segment definitions ---
-
-const RUN_CHART_SEGMENTS = [
-  { key: 'success', cls: 'usage-chart-bar-success' },
-  { key: 'error', cls: 'usage-chart-bar-error' },
-  { key: 'running', cls: 'usage-chart-bar-running' },
-];
-
-const TOKEN_CHART_SEGMENTS = [
-  { key: 'input', cls: 'usage-chart-bar-running' },
-  { key: 'output', cls: 'usage-chart-bar-success' },
-];
-
-// --- Helpers ---
-
-function _td(text, attrs = {}) {
-  return _el('td', { ...attrs, textContent: text });
-}
-
-/**
- * Build a <tr> from an array of column descriptors.
+ * Public entry point for UsageView tab configs.
  *
- * Each entry can be:
- *   - A DOM Node (inserted as-is into the row)
- *   - An object { value, className?, style?, title? } → converted via _td()
- *
- * @param {Array<Node | { value: string|number, className?: string, style?: Record<string, string>, title?: string }>} columns
- * @returns {HTMLTableRowElement}
- */
-function buildTableRow(columns) {
-  const cells = columns.map((col) => {
-    if (col instanceof Node) return col;
-    const attrs = {};
-    if (col.className) attrs.className = col.className;
-    if (col.style) attrs.style = col.style;
-    if (col.title) attrs.title = col.title;
-    return _td(col.value, attrs);
-  });
-  return _el('tr', {}, ...cells);
-}
-
-function tokenTooltip(day) {
-  return `${day.label}: ${formatTokens(day.total)} (in: ${formatTokens(day.input)}, out: ${formatTokens(day.output)})`;
-}
-
-// --- Pure DOM builders ---
-
-function createBarCell(pct) {
-  return _el('td', { className: 'usage-file-bar-cell' },
-    _el('div', { className: 'usage-file-bar' },
-      _el('div', { className: 'usage-file-bar-fill', style: { width: `${pct}%` } }),
-    ),
-  );
-}
-
-export function createSection(title) {
-  return _el('div', { className: 'usage-section' },
-    _el('div', { className: 'usage-section-title', textContent: title }),
-  );
-}
-
-// --- Tab config typedefs ---
-
-/**
- * A per-tab metrics slice extracted from the full metrics object. Its shape
- * varies per tab (agent / flow / token), so it is kept as a loose record.
- * @typedef {Record<string, unknown>} MetricsSlice
+ * Implementation is split across:
+ *   - usage-view-shared.js     : constants, pure helpers, factory
+ *   - usage-view-agents-tab.js : agents tab
+ *   - usage-view-tokens-tab.js : tokens tab
+ *   - usage-view-flows-tab.js  : flows tab
  */
 
-/**
- * @typedef {object} UsageCard
- * @property {string} label         - card label
- * @property {string|number} value  - card value
- * @property {string} cls           - CSS class applied to the value
- * @property {string} [sub]         - optional sub-text shown below the value
- */
+import { agentTabConfig } from './usage-view-agents-tab.js';
+import { tokenTabConfig } from './usage-view-tokens-tab.js';
+import { flowTabConfig } from './usage-view-flows-tab.js';
 
-/**
- * @typedef {object} UsageChart
- * @property {string} title                                 - chart title
- * @property {Array<Record<string, unknown>>} data          - per-day data points
- * @property {Array<{ key: string, cls: string }>} segments - stacked-bar segment definitions
- * @property {(day: Record<string, unknown>) => string} tooltip - builds the tooltip text for a day
- */
-
-/**
- * @typedef {object} UsageTable
- * @property {string} title      - table title
- * @property {string[]} headers  - column headers
- * @property {string} tableCls   - CSS class applied to the <table>
- * @property {Array<Record<string, unknown>>} data - row data
- * @property {(row: Record<string, unknown>) => HTMLTableRowElement} renderRow - renders one <tr>
- */
-
-/**
- * A built tab configuration.
- * @typedef {{ cards: UsageCard[], chart: UsageChart, tables: UsageTable[] }} TabConfig
- */
-
-/**
- * An empty-state descriptor returned in place of a TabConfig.
- * @typedef {{ empty: string[] }} EmptyTabConfig
- */
-
-// --- Tab configurations ---
-
-/**
- * Build the two run-metric cards shared by agents and flows tabs:
- * success rate + average duration (with min/max sub-text).
- */
-function _runMetricCards(m) {
-  return [
-    { label: 'Taux succès', value: `${m.rate.rate}%`, cls: rateCls(m.rate.rate) },
-    { label: 'Durée moy.', value: formatDuration(m.duration.avg), cls: 'usage-stat-value-blue', sub: m.duration.count > 0 ? `min: ${formatDuration(m.duration.min)} · max: ${formatDuration(m.duration.max)}` : '' },
-  ];
-}
-
-/**
- * Options object describing the tab-specific parts of a run-based tab.
- * Consumed by {@link _createRunBasedTabConfig}.
- *
- * @typedef {object} RunBasedTabOptions
- * @property {string} sliceKey - Key to extract the metrics slice (e.g. 'agent', 'flow').
- * @property {(m: MetricsSlice) => UsageCard[]} headerCards - Returns the first 2 cards specific to this tab.
- * @property {string} chartTitle - Title displayed above the chart.
- * @property {(m: MetricsSlice, metrics: Record<string, MetricsSlice>) => UsageTable[]} tables - Returns table descriptor(s).
- * @property {(m: MetricsSlice) => {empty: string[]}|null} [emptyGuard] - Optional early-return for empty state.
- */
-
-/**
- * Factory that builds the common { cards, chart, tables } shape shared by
- * run-based tabs (agents and flows).
- *
- * Both run-based tabs follow the same pattern:
- *  1. Extract a metrics slice from the top-level metrics object.
- *  2. Build 2 tab-specific "header" cards, then append the 2 shared run-metric
- *     cards (success rate + avg duration) via _runMetricCards().
- *  3. Create a daily-run chart with RUN_CHART_SEGMENTS.
- *  4. Attach one or more data tables.
- *
- * Callers only supply the parts that differ (sliceKey, headerCards, chartTitle,
- * tables builder) — everything else is handled here.
- *
- * @param {Record<string, MetricsSlice>}   metrics  The full metrics object.
- * @param {RunBasedTabOptions} options  Tab-specific configuration.
- * @returns {TabConfig | EmptyTabConfig}
- */
-function _createRunBasedTabConfig(metrics, { sliceKey, headerCards, chartTitle, tables, emptyGuard }) {
-  const m = metrics[sliceKey];
-
-  if (emptyGuard) {
-    const guard = emptyGuard(m);
-    if (guard) return guard;
-  }
-
-  return {
-    cards: [...headerCards(m), ..._runMetricCards(m)],
-    chart: { title: chartTitle, data: m.perDay, segments: RUN_CHART_SEGMENTS, tooltip: runTooltip },
-    tables: tables(m, metrics),
-  };
-}
+export { TABS, createSection, _internals } from './usage-view-shared.js';
 
 export function getTabConfig(tabId, metrics) {
-  const builders = { agents: _agentTabConfig, tokens: _tokenTabConfig, flows: _flowTabConfig };
+  const builders = { agents: agentTabConfig, tokens: tokenTabConfig, flows: flowTabConfig };
   return builders[tabId]?.(metrics);
-}
-
-function _renderAgentRow(a) {
-  return buildTableRow([
-    { value: a.agent, className: 'usage-flow-name' },
-    { value: a.totalSessions },
-    { value: a.active, style: { color: a.active > 0 ? 'var(--green)' : 'var(--text-muted)' } },
-    { value: `${a.successRate}%`, className: 'usage-flow-rate', style: { color: rateColor(a.successRate) } },
-    { value: a.avgDuration > 0 ? formatDuration(a.avgDuration) : '\u2014', className: 'usage-flow-duration' },
-  ]);
-}
-
-function _renderFileRow(maxFileCount) {
-  return (file) => buildTableRow([
-    { value: file.file, className: 'usage-file-name', title: file.file },
-    { value: file.count, className: 'usage-file-count' },
-    createBarCell((file.count / maxFileCount) * 100),
-  ]);
-}
-
-function _agentHeaderCards(m) {
-  return [
-    { label: 'Sessions', value: m.totalSessions, cls: '' },
-    { label: 'En cours', value: m.activeSessions, cls: m.activeSessions > 0 ? 'usage-stat-value-green' : '' },
-  ];
-}
-
-function _agentTabConfig(metrics) {
-  return _createRunBasedTabConfig(metrics, {
-    sliceKey: 'agent',
-    headerCards: _agentHeaderCards,
-    chartTitle: 'Sessions par jour',
-    tables: (m, allMetrics) => {
-      const maxFileCount = allMetrics.mostModifiedFiles[0]?.count || 1;
-      return [
-        { title: 'Par agent', headers: ['Agent', 'Sessions', 'Actifs', 'Succès', 'Durée moy.'], tableCls: 'usage-flow-table', data: m.byAgent, renderRow: _renderAgentRow },
-        { title: 'Fichiers les plus modifiés (30 jours)', headers: ['Fichier', 'Modifs', ''], tableCls: 'usage-files-table', data: allMetrics.mostModifiedFiles, renderRow: _renderFileRow(maxFileCount) },
-      ];
-    },
-  });
-}
-
-function _tokenCards(t) {
-  return [
-    { label: 'Total', value: formatTokens(t.total), cls: '' },
-    { label: 'Input', value: formatTokens(t.totalInput), cls: 'usage-stat-value-blue' },
-    { label: 'Output', value: formatTokens(t.totalOutput), cls: 'usage-stat-value-green' },
-    { label: 'Cache read', value: formatTokens(t.totalCacheRead), cls: '', sub: t.totalCacheCreate > 0 ? `cache write: ${formatTokens(t.totalCacheCreate)}` : '' },
-  ];
-}
-
-function _renderProjectRow(maxProjectTotal) {
-  return (proj) => buildTableRow([
-    { value: proj.project, className: 'usage-file-name' },
-    { value: formatTokens(proj.input), className: 'usage-file-count', style: { color: 'var(--blue)' } },
-    { value: formatTokens(proj.output), className: 'usage-file-count', style: { color: 'var(--green)' } },
-    { value: formatTokens(proj.total), className: 'usage-file-count' },
-    createBarCell((proj.total / maxProjectTotal) * 100),
-  ]);
-}
-
-function _tokenTabConfig(metrics) {
-  const t = metrics.tokens;
-  if (!t || t.total === 0) {
-    return { empty: ['Aucune donnée de tokens', 'Les tokens sont lus depuis les sessions Claude (~/.claude/projects/)'] };
-  }
-  const maxProjectTotal = t.perProject?.[0]?.total || 1;
-  return {
-    cards: _tokenCards(t),
-    chart: { title: 'Tokens par jour (30 derniers jours)', data: t.perDay, segments: TOKEN_CHART_SEGMENTS, tooltip: tokenTooltip },
-    tables: [
-      { title: 'Par projet', headers: ['Projet', 'Input', 'Output', 'Total', ''], tableCls: 'usage-files-table', data: t.perProject, renderRow: _renderProjectRow(maxProjectTotal) },
-    ],
-  };
-}
-
-/** @internal Exposed for unit tests only. */
-export const _internals = { buildTableRow };
-
-function _flowNameCell(flow) {
-  return _el('td', {},
-    _el('span', {
-      className: `usage-flow-name ${!flow.enabled ? 'usage-flow-disabled' : ''}`,
-      textContent: flow.name,
-    }),
-    !flow.enabled && _el('span', {
-      style: { fontSize: '10px', color: 'var(--text-muted)', marginLeft: '6px' },
-      textContent: '(désactivé)',
-    }),
-  );
-}
-
-function _renderFlowRow(flow) {
-  return buildTableRow([
-    _flowNameCell(flow),
-    { value: flow.totalRuns },
-    { value: `${flow.successRate}%`, className: 'usage-flow-rate', style: { color: rateColor(flow.successRate) } },
-    { value: flow.avgDuration > 0 ? formatDuration(flow.avgDuration) : '\u2014', className: 'usage-flow-duration' },
-  ]);
-}
-
-function _flowHeaderCards(f) {
-  return [
-    { label: 'Total Runs', value: f.rate.total, cls: '' },
-    { label: 'Flows actifs', value: `${f.activeFlows}/${f.totalFlows}`, cls: '' },
-  ];
-}
-
-function _flowTabConfig(metrics) {
-  return _createRunBasedTabConfig(metrics, {
-    sliceKey: 'flow',
-    headerCards: _flowHeaderCards,
-    chartTitle: 'Runs par jour',
-    emptyGuard: (f) => f.totalFlows === 0
-      ? { empty: ['Aucun flow configuré', 'Créez des flows depuis la vue FLOW'] }
-      : null,
-    tables: (f) => [
-      { title: 'Par flow', headers: ['Flow', 'Runs', 'Succès', 'Durée moy.'], tableCls: 'usage-flow-table', data: f.flowStats, renderRow: _renderFlowRow },
-    ],
-  });
 }
